move hook to the new folder
