+++ graylog2-web-interface/src/brand-customization/useCustomLogo.ts
